mqtt消息接受从市级开始

+++ app/src/main/java/lilun/com/pension/ui/help/help_detail/AskDetailFragment.java
@@ -213,7 +213,7 @@ public class AskDetailFragment extends BaseFragment<HelpDetailContract.Presenter
 
 
         //显示补贴
-        int price = mAid.getPrice();
+        long price = mAid.getPrice();
         if (price == 0) {
             tvPrice.setVisibility(View.INVISIBLE);
         } else {
